Docbrief-main/requirements.txt
Flask
Werkzeug
google-generativeai
PyPDF2
python-docx
